+++ app/src/main/java/activities/MentorDetail.java
@@ -1,17 +1,14 @@
 package activities;
 
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.wittenPortfolio.R;
 
@@ -54,42 +51,32 @@ public class MentorDetail extends AppCompatActivity {
 
     private void deleteSetup() {
         deleteMentorBtn = findViewById(R.id.deleteMentorBTN);
-        deleteMentorBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //Confirm the delete with the user.
-                AlertDialog.Builder builder = new AlertDialog.Builder(MentorDetail.this);
-                builder.setCancelable(true);
-                builder.setTitle("Are You Sure?");
-                builder.setMessage("Deleting a mentor will delete all of their associated classes, this cannot be undone\n" +
-                        "\n" +
-                        "We recommend updating class information before deleting a mentor.\n" +
-                        "\n" +
-                        "Would you like to continue?");
-                builder.setPositiveButton("Confirm",
-                        new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which) {
-                                //Delete associated courses.
-                                db.courseDAO().deleteMentorCourses(currentMentor.id);
-                                //Delete the Mentor
-                                db.mentorDAO().delete(currentMentor);
-                                //Return to mentor list
-                                Intent intent = new Intent(MentorDetail.this, MentorList.class);
-                                startActivity(intent);
-                            }
-                        });
-                builder.setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        //Return without doing anything.
-                        return;
-                    }
-                });
-                AlertDialog dialog = builder.create();
-                dialog.show();
+        deleteMentorBtn.setOnClickListener(v -> {
+            //Confirm the delete with the user.
+            AlertDialog.Builder builder = new AlertDialog.Builder(MentorDetail.this);
+            builder.setCancelable(true);
+            builder.setTitle("Are You Sure?");
+            builder.setMessage("Deleting a mentor will delete all of their associated classes, this cannot be undone\n" +
+                    "\n" +
+                    "We recommend updating class information before deleting a mentor.\n" +
+                    "\n" +
+                    "Would you like to continue?");
+            builder.setPositiveButton("Confirm",
+                    (dialog, which) -> {
+                        //Delete associated courses.
+                        db.courseDAO().deleteMentorCourses(currentMentor.id);
+                        //Delete the Mentor
+                        db.mentorDAO().delete(currentMentor);
+                        //Return to mentor list
+                        Intent intent = new Intent(MentorDetail.this, MentorList.class);
+                        startActivity(intent);
+                    });
+            builder.setNegativeButton(android.R.string.cancel, (dialog, which) -> {
+                //Return without doing anything.
+            });
+            AlertDialog dialog = builder.create();
+            dialog.show();
 
-            }
         });
     }
 
@@ -99,12 +86,12 @@ public class MentorDetail extends AppCompatActivity {
         courseList = db.courseDAO().getAllCourses();
         //Find associated courses.
         for(Course c: courseList){
-            if(c.mentorID == currentMentor.id){
+            if(Objects.equals(c.mentorID, currentMentor.id)){
                 classNames.add(c.courseName);
             }
         }
         //Define an adapter
-        ArrayAdapter<String> adapter =new ArrayAdapter<String>
+        ArrayAdapter<String> adapter = new ArrayAdapter<>
                 (this, R.layout.course_name_row, classNames);
         listView.setAdapter(adapter);
     }
